moved account queries into authslice
