+++ src/features/auth/authApiSlice.js
@@ -8,11 +8,13 @@ export const authApiSlice = apiSlice.injectEndpoints({
         method: 'POST',
         body: { ...credentials },
       }),
+      providesTags: ['User'],
     }),
     logout: builder.mutation({
       query: () => ({
         url: '/logout',
       }),
+      invalidatesTags: ['User'],
     }),
     registerAccount: builder.mutation({
       query: (register) => ({
@@ -21,6 +23,44 @@ export const authApiSlice = apiSlice.injectEndpoints({
         body: register,
       }),
     }),
+    changeUserEmail: builder.mutation({
+      query: (email) => ({
+        url: '/email',
+        method: 'PUT',
+        body: email,
+      }),
+      invalidatesTags: ['User'],
+    }),
+    changeUserPassword: builder.mutation({
+      query: (password) => ({
+        url: '/password',
+        method: 'PUT',
+        body: password,
+      }),
+      invalidatesTags: ['User'],
+    }),
+    // ! TODO
+    deleteUser: builder.mutation({
+      query: () => ({
+        url: '/api/user',
+        method: 'DELETE',
+      }),
+      invalidatesTags: ['User'],
+    }),
+    validateToken: builder.query({
+      query: (req) => `/validation/${req.id}/${req.token}`,
+      transformResponse: (responseData) => {
+        return responseData;
+      },
+    }),
+    resetUserPassword: builder.mutation({
+      query: ({ id, token, newPassword }) => ({
+        url: `/reset-pw/${id}/${token}`,
+        method: 'PUT',
+        body: { newPassword },
+      }),
+      invalidatesTags: ['User'],
+    }),
   }),
 });
 
@@ -28,4 +68,9 @@ export const {
   useLoginMutation,
   useLogoutMutation,
   useRegisterAccountMutation,
+  useChangeUserEmailMutation,
+  useChangeUserPasswordMutation,
+  useDeleteUserMutation,
+  useResetUserPasswordMutation,
+  useValidateTokenQuery,
 } = authApiSlice;
