+++ src/hooks/useAuth.js
@@ -3,12 +3,17 @@ import { useNavigate, useLocation } from 'react-router-dom';
 import {
   useLoginMutation,
   useLogoutMutation,
+  useChangeUserEmailMutation,
+  useChangeUserPasswordMutation,
+  useDeleteUserMutation,
+  useResetUserPasswordMutation,
 } from '@/features/auth/authApiSlice';
 import {
   logoutUser,
   selectCurrentUser,
   selectCurrentUserRole,
   setCredentials,
+  updateUserDetails,
 } from '@/features/auth/authSlice';
 import { useNotification } from './useNotification';
 import { cleanEmail } from '@/utils/email';
@@ -20,6 +25,10 @@ export function useAuth() {
   const role = useSelector(selectCurrentUserRole);
   const [login] = useLoginMutation();
   const [logout] = useLogoutMutation();
+  const [changeUserEmail] = useChangeUserEmailMutation();
+  const [changeUserPassword] = useChangeUserPasswordMutation();
+  const [deleteUser] = useDeleteUserMutation();
+  const [resetUserPassword] = useResetUserPasswordMutation();
 
   const location = useLocation();
 
@@ -56,5 +65,68 @@ export function useAuth() {
     }
   };
 
-  return { user, role, handleLogin, handleLogout };
+  const handleUserEmailChange = async (newEmailObj) => {
+    try {
+      const updatedUser = await changeUserEmail(newEmailObj).unwrap();
+      if (updatedUser.success) {
+        dispatch(
+          updateUserDetails({
+            user: updatedUser.user.email,
+            role: updatedUser.user.role,
+          })
+        );
+        handleSuccess(updatedUser.message);
+        return true;
+      }
+    } catch (error) {
+      handleError(`Error changing email: ${error}`);
+    }
+  };
+
+  const handleUserPasswordChange = async (newPasswordObj) => {
+    try {
+      const updatedUser = await changeUserPassword(newPasswordObj).unwrap();
+      if (updatedUser.success) {
+        handleSuccess(updatedUser.message);
+        return true;
+      }
+    } catch (error) {
+      handleError(`Error changing password: ${error}`);
+    }
+  };
+
+  const handleUserDelete = async () => {
+    try {
+      const deletedUser = await deleteUser().unwrap();
+      if (deletedUser.success) {
+        handleSuccess(deletedUser.message);
+        dispatch(logoutUser());
+      }
+    } catch (error) {
+      handleError(`Error deleting user: ${error}`);
+    }
+  };
+
+  const handleUserPasswordReset = async (newCredentials) => {
+    try {
+      const updatedUser = await resetUserPassword(newCredentials).unwrap();
+      if (updatedUser.success) {
+        handleSuccess(updatedUser.message);
+        return true;
+      }
+    } catch (err) {
+      handleError(err);
+    }
+  };
+
+  return {
+    user,
+    role,
+    handleLogin,
+    handleLogout,
+    handleUserEmailChange,
+    handleUserPasswordChange,
+    handleUserDelete,
+    handleUserPasswordReset,
+  };
 }
